requirements.txt
flask
flask-cors
ultralytics
pillow
numpy
requests
pyyaml
torch
torchvision